Rename all the files for Azure Key Vault
