+++ src/attributes.h
@@ -1,5 +1,5 @@
-#include "pkcs11_compat.h"
-#include "aws_kms_slot.h"
+#include "pkcs11-compat.h"
+#include "azure-keyvault-slot.h"
 
 CK_RV getKmsKeyAttributeValue(AwsKmsSlot& slot, CK_ATTRIBUTE_TYPE attr, CK_VOID_PTR pValue, CK_ULONG_PTR pulValueLen);
 CK_RV getCertificateAttributeValue(AwsKmsSlot& slot, CK_ATTRIBUTE_TYPE attr, CK_VOID_PTR pValue, CK_ULONG_PTR pulValueLen);
