+++ src/azure-keyvault-slot.h
